+++ src/all/batotov4/AndroidManifest.xml
@@ -2,7 +2,7 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android">
 <application>
     <activity
-        android:name=".all.batotov4.BatoToV4UrlActivity"
+        android:name=".all.batotov4.UrlActivity"
         android:excludeFromRecents="true"
         android:exported="true"
         android:theme="@android:style/Theme.NoDisplay">
@@ -12,16 +12,38 @@
             <category android:name="android.intent.category.DEFAULT" />
             <category android:name="android.intent.category.BROWSABLE" />
 
-            <data android:host="*.bato.si" />
             <data android:host="bato.si" />
-            <data android:host="*.bato.ing" />
             <data android:host="bato.ing" />
+            <data android:host="bato.to" />
+            <data android:host="batocomic.com" />
+            <data android:host="batocomic.net" />
+            <data android:host="batocomic.org" />
+            <data android:host="batotoo.com" />
+            <data android:host="batotwo.com" />
+            <data android:host="battwo.com" />
+            <data android:host="comiko.net" />
+            <data android:host="comiko.org" />
+            <data android:host="mangatoto.com" />
+            <data android:host="mangatoto.net" />
+            <data android:host="mangatoto.org" />
+            <data android:host="readtoto.com" />
+            <data android:host="readtoto.net" />
+            <data android:host="readtoto.org" />
+            <data android:host="xbato.com" />
+            <data android:host="xbato.net" />
+            <data android:host="xbato.org" />
+            <data android:host="zbato.com" />
+            <data android:host="zbato.net" />
+            <data android:host="zbato.org" />
+            <data android:host="dto.to" />
+            <data android:host="fto.to" />
+            <data android:host="hto.to" />
+            <data android:host="jto.to" />
+            <data android:host="mto.to" />
+            <data android:host="wto.to" />
 
             <data
                 android:pathPattern="/series/..*"
-                android:scheme="https" />
-            <data
-                android:pathPattern="/subject-overview/..*"
                 android:scheme="https" />
             <data
                 android:pathPattern="/title/..*"
