--- src/all/batotov4/CHANGELOG.md
@@ -1,6 +1,0 @@
-## 1.4.1
-
-### Features
-
-* Conversion from Bato.to to Bato.to V4
-* First version
