--- src/all/batotov4/README.md
@@ -1,20 +1,0 @@
-# Bato.to
-
-Table of Content
-- [FAQ](#FAQ)
-    - [Why are there Manga of diffrent languge than the selected one in Personal & Utils lists?](#why-are-there-manga-of-diffrent-languge-than-the-selected-one-in-personal--utils-lists)
-    - [Bato.to is not loading anything?](#batoto-is-not-loading-anything)
-
-[Uncomment this if needed; and replace &#40; and &#41; with ( and )]: <> (- [Guides]&#40;#Guides&#41;)
-
-Don't find the question you are look for go check out our general FAQs and Guides over at [Extension FAQ](https://tachiyomi.org/help/faq/#extensions) or [Getting Started](https://tachiyomi.org/help/guides/getting-started/#installation)
-
-## FAQ
-
-### Why are there Manga of diffrent languge than the selected one in Personal & Utils lists?
-Personol & Utils lists have no way to difritiate between langueges.
-
-### Bato.to is not loading anything?
-Bato.to get blocked by some ISPs, try using a diffrent mirror of Bato.to from the settings.
-
-[Uncomment this if needed]: <> (## Guides)
