+++ src/all/batotov4/build.gradle
@@ -6,7 +6,3 @@ ext {
 }
 
 apply from: "$rootDir/common.gradle"
-
-dependencies {
-    implementation(project(':lib:cryptoaes'))
-}
