+++ src/all/batotov4/src/eu/kanade/tachiyomi/extension/all/batotov4/BatoToV4Factory.kt
@@ -1,116 +1,112 @@
 package eu.kanade.tachiyomi.extension.all.batotov4
 
-import eu.kanade.tachiyomi.source.Source
 import eu.kanade.tachiyomi.source.SourceFactory
 
 class BatoToV4Factory : SourceFactory {
-    override fun createSources(): List<Source> = languages.map { BatoToV4(it.lang, it.siteLang) }
+    override fun createSources() = listOf(
+        BatoToV4("all", ""),
+        BatoToV4("en"),
+        BatoToV4("ar"),
+        BatoToV4("bg"),
+        BatoToV4("zh"),
+        BatoToV4("cs"),
+        BatoToV4("da"),
+        BatoToV4("nl"),
+        BatoToV4("fil"),
+        BatoToV4("fi"),
+        BatoToV4("fr"),
+        BatoToV4("de"),
+        BatoToV4("el"),
+        BatoToV4("he"),
+        BatoToV4("hi"),
+        BatoToV4("hu"),
+        BatoToV4("id"),
+        BatoToV4("it"),
+        BatoToV4("ja"),
+        BatoToV4("ko"),
+        BatoToV4("ms"),
+        BatoToV4("pl"),
+        BatoToV4("pt"),
+        BatoToV4("pt-BR", "pt_br"),
+        BatoToV4("ro"),
+        BatoToV4("ru"),
+        BatoToV4("es"),
+        BatoToV4("es-419", "es_419"),
+        BatoToV4("sv"),
+        BatoToV4("th"),
+        BatoToV4("tr"),
+        BatoToV4("uk"),
+        BatoToV4("vi"),
+        BatoToV4("af"),
+        BatoToV4("sq"),
+        BatoToV4("am"),
+        BatoToV4("hy"),
+        BatoToV4("az"),
+        BatoToV4("be"),
+        BatoToV4("bn"),
+        BatoToV4("bs"),
+        BatoToV4("my"),
+        BatoToV4("km"),
+        BatoToV4("ca"),
+        BatoToV4("ceb"),
+        BatoToV4("zh-Hans", "zh_hk"),
+        BatoToV4("zh-Hant", "zh_tw"),
+        BatoToV4("hr"),
+        BatoToV4("et"),
+        BatoToV4("fo"),
+        BatoToV4("ka"),
+        BatoToV4("gn"),
+        BatoToV4("gu"),
+        BatoToV4("ht"),
+        BatoToV4("ha"),
+        BatoToV4("is"),
+        BatoToV4("ig"),
+        BatoToV4("ga"),
+        BatoToV4("jv"),
+        BatoToV4("kn"),
+        BatoToV4("kk"),
+        BatoToV4("ku"),
+        BatoToV4("ky"),
+        BatoToV4("lo"),
+        BatoToV4("lv"),
+        BatoToV4("lt"),
+        BatoToV4("lb"),
+        BatoToV4("mk"),
+        BatoToV4("mg"),
+        BatoToV4("ml"),
+        BatoToV4("mt"),
+        BatoToV4("mi"),
+        BatoToV4("mr"),
+        BatoToV4("mo"),
+        BatoToV4("mn"),
+        BatoToV4("ne"),
+        BatoToV4("no"),
+        BatoToV4("ny"),
+        BatoToV4("ps"),
+        BatoToV4("fa"),
+        BatoToV4("rm"),
+        BatoToV4("sm"),
+        BatoToV4("sr"),
+        BatoToV4("sh"),
+        BatoToV4("st"),
+        BatoToV4("sn"),
+        BatoToV4("sd"),
+        BatoToV4("si"),
+        BatoToV4("sk"),
+        BatoToV4("sl"),
+        BatoToV4("so"),
+        BatoToV4("sw"),
+        BatoToV4("tg"),
+        BatoToV4("ta"),
+        BatoToV4("te"),
+        BatoToV4("th"),
+        BatoToV4("ti"),
+        BatoToV4("to"),
+        BatoToV4("tk"),
+        BatoToV4("ur"),
+        BatoToV4("uz"),
+        BatoToV4("yo"),
+        BatoToV4("zu"),
+        BatoToV4("other", "_t"),
+    )
 }
-
-class LanguageOption(val lang: String, val siteLang: String = lang)
-private val languages = listOf(
-    LanguageOption("all", ""),
-    LanguageOption("en"),
-    LanguageOption("ar"),
-    LanguageOption("bg"),
-    LanguageOption("zh"),
-    LanguageOption("cs"),
-    LanguageOption("da"),
-    LanguageOption("nl"),
-    LanguageOption("fil"),
-    LanguageOption("fi"),
-    LanguageOption("fr"),
-    LanguageOption("de"),
-    LanguageOption("el"),
-    LanguageOption("he"),
-    LanguageOption("hi"),
-    LanguageOption("hu"),
-    LanguageOption("id"),
-    LanguageOption("it"),
-    LanguageOption("ja"),
-    LanguageOption("ko"),
-    LanguageOption("ms"),
-    LanguageOption("pl"),
-    LanguageOption("pt"),
-    LanguageOption("pt-BR", "pt_br"),
-    LanguageOption("ro"),
-    LanguageOption("ru"),
-    LanguageOption("es"),
-    LanguageOption("es-419", "es_419"),
-    LanguageOption("sv"),
-    LanguageOption("th"),
-    LanguageOption("tr"),
-    LanguageOption("uk"),
-    LanguageOption("vi"),
-    LanguageOption("af"),
-    LanguageOption("sq"),
-    LanguageOption("am"),
-    LanguageOption("hy"),
-    LanguageOption("az"),
-    LanguageOption("be"),
-    LanguageOption("bn"),
-    LanguageOption("bs"),
-    LanguageOption("my"),
-    LanguageOption("km"),
-    LanguageOption("ca"),
-    LanguageOption("ceb"),
-    LanguageOption("zh-Hans", "zh_hk"),
-    LanguageOption("zh-Hant", "zh_tw"),
-    LanguageOption("hr"),
-    LanguageOption("et"),
-    LanguageOption("fo"),
-    LanguageOption("ka"),
-    LanguageOption("gn"),
-    LanguageOption("gu"),
-    LanguageOption("ht"),
-    LanguageOption("ha"),
-    LanguageOption("is"),
-    LanguageOption("ig"),
-    LanguageOption("ga"),
-    LanguageOption("jv"),
-    LanguageOption("kn"),
-    LanguageOption("kk"),
-    LanguageOption("ku"),
-    LanguageOption("ky"),
-    LanguageOption("lo"),
-    LanguageOption("lv"),
-    LanguageOption("lt"),
-    LanguageOption("lb"),
-    LanguageOption("mk"),
-    LanguageOption("mg"),
-    LanguageOption("ml"),
-    LanguageOption("mt"),
-    LanguageOption("mi"),
-    LanguageOption("mr"),
-    LanguageOption("mo"),
-    LanguageOption("mn"),
-    LanguageOption("ne"),
-    LanguageOption("no"),
-    LanguageOption("ny"),
-    LanguageOption("ps"),
-    LanguageOption("fa"),
-    LanguageOption("rm"),
-    LanguageOption("sm"),
-    LanguageOption("sr"),
-    LanguageOption("sh"),
-    LanguageOption("st"),
-    LanguageOption("sn"),
-    LanguageOption("sd"),
-    LanguageOption("si"),
-    LanguageOption("sk"),
-    LanguageOption("sl"),
-    LanguageOption("so"),
-    LanguageOption("sw"),
-    LanguageOption("tg"),
-    LanguageOption("ta"),
-    LanguageOption("te"),
-    LanguageOption("th"),
-    LanguageOption("ti"),
-    LanguageOption("to"),
-    LanguageOption("tk"),
-    LanguageOption("ur"),
-    LanguageOption("uz"),
-    LanguageOption("yo"),
-    LanguageOption("zu"),
-    LanguageOption("other", "_t"),
-)
