+++ src/all/batotov4/src/eu/kanade/tachiyomi/extension/all/batotov4/Queries.kt
@@ -1,13 +1,29 @@
 package eu.kanade.tachiyomi.extension.all.batotov4
 
-import eu.kanade.tachiyomi.extension.all.batotov4.BatoToV4.Companion.whitespace
-
 private fun buildQuery(queryAction: () -> String): String {
     return queryAction()
-        .trimIndent()
         .replace("%", "$")
         .replace(whitespace, " ")
 }
+
+private val whitespace = Regex("\\s+")
+
+private const val COMIC_NODE = """
+    data {
+        id
+        name
+        altNames
+        authors
+        artists
+        originalStatus
+        uploadStatus
+        genres
+        summary
+        extraInfo
+        urlPath
+        urlCoverOri
+    }
+"""
 
 // Query for comic search
 val COMIC_SEARCH_QUERY: String = buildQuery {
@@ -15,20 +31,10 @@ val COMIC_SEARCH_QUERY: String = buildQuery {
         query (%select: Comic_Browse_Select) {
             get_comic_browse(select: %select) {
                 paging {
-                    pages
-                    page
                     next
                 }
                 items {
-                    data {
-                        id
-                        name
-                        urlPath
-                        urlCover300
-                        urlCover600
-                        urlCover900
-                        urlCoverOri
-                    }
+                    $COMIC_NODE
                 }
             }
         }
@@ -40,23 +46,7 @@ val COMIC_NODE_QUERY: String = buildQuery {
     """
         query get_comicNode(%id: ID!) {
             get_comicNode(id: %id) {
-                data {
-                    id
-                    name
-                    altNames
-                    authors
-                    artists
-                    originalStatus
-                    uploadStatus
-                    genres
-                    summary
-                    extraInfo
-                    urlPath
-                    urlCover300
-                    urlCover600
-                    urlCover900
-                    urlCoverOri
-                }
+                $COMIC_NODE
             }
         }
     """
@@ -68,13 +58,18 @@ val CHAPTER_LIST_QUERY: String = buildQuery {
         query get_comic_chapterList(%comicId: ID!, %start: Int) {
             get_comic_chapterList(comicId: %comicId, start: %start) {
                 data {
+                    comicId
                     id
                     dname
                     title
-                    urlPath
                     dateCreate
                     dateModify
                     userNode {
+                        data {
+                            name
+                        }
+                    }
+                    groupNodes {
                         data {
                             name
                         }
@@ -91,6 +86,8 @@ val CHAPTER_NODE_QUERY: String = buildQuery {
         query get_chapterNode(%id: ID!) {
             get_chapterNode(id: %id) {
                 data {
+                    id,
+                    comicId,
                     imageFile {
                         urlList
                     }
@@ -106,20 +103,10 @@ val SSER_MY_UPDATES_QUERY: String = buildQuery {
         query (%select: Sser_MyUpdates_Select) {
             get_sser_myUpdates(select: %select) {
                 paging {
-                    pages
-                    page
                     next
                 }
                 items {
-                    data {
-                        id
-                        name
-                        urlPath
-                        urlCover300
-                        urlCover600
-                        urlCover900
-                        urlCoverOri
-                    }
+                    $COMIC_NODE
                 }
             }
         }
@@ -135,15 +122,7 @@ val SSER_MY_HISTORY_QUERY: String = buildQuery {
                 newStart
                 items {
                     comicNode {
-                        data {
-                            id
-                            name
-                            urlPath
-                            urlCover300
-                            urlCover600
-                            urlCover900
-                            urlCoverOri
-                        }
+                        $COMIC_NODE
                     }
                 }
             }
@@ -157,20 +136,10 @@ val USER_COMIC_LIST_QUERY: String = buildQuery {
         query (%select: User_ComicList_Select) {
             get_user_comicList(select: %select) {
                 paging {
-                    pages
-                    page
                     next
                 }
                 items {
-                    data {
-                        id
-                        name
-                        urlPath
-                        urlCover300
-                        urlCover600
-                        urlCover900
-                        urlCoverOri
-                    }
+                    $COMIC_NODE
                 }
             }
         }
