+++ src/all/batotov4/src/eu/kanade/tachiyomi/extension/all/batotov4/UrlActivity.kt
@@ -1,0 +1,29 @@
+package eu.kanade.tachiyomi.extension.all.batotov4
+
+import android.app.Activity
+import android.content.ActivityNotFoundException
+import android.content.Intent
+import android.os.Bundle
+import android.util.Log
+import kotlin.system.exitProcess
+
+class UrlActivity : Activity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+
+        val mainIntent = Intent().apply {
+            action = "eu.kanade.tachiyomi.SEARCH"
+            putExtra("query", intent.data.toString())
+            putExtra("filter", packageName)
+        }
+
+        try {
+            startActivity(mainIntent)
+        } catch (e: ActivityNotFoundException) {
+            Log.e("BatoTo v4", "Unable to launch activity", e)
+        }
+
+        finish()
+        exitProcess(0)
+    }
+}
